refactor: auth module clean up and compose navigation with root navigation
