+++ app/src/main/java/org/mifos/mobile/ui/activities/GuarantorActivity.kt
@@ -1,34 +1,33 @@
-package org.mifos.mobile.ui.login
+package org.mifos.mobile.ui.activities
 
-import android.content.Intent
 import android.os.Bundle
+import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.navigation.compose.rememberNavController
+import org.mifos.mobile.feature.guarantor.navigation.GuarantorScreen
+import org.mifos.mobile.feature.guarantor.navigation.SetUpGuarantorNavGraph
 import dagger.hilt.android.AndroidEntryPoint
+import org.mifos.mobile.core.common.Constants.LOAN_ID
 import org.mifos.mobile.core.ui.theme.MifosMobileTheme
-import org.mifos.mobile.feature.auth.navigation.AuthenticationRoute
-import org.mifos.mobile.navigation.RootNavGraph
-import org.mifos.mobile.ui.activities.base.BaseActivity
 
-/**
- * @author Vishwajeet
- * @since 05/06/16
- */
 @AndroidEntryPoint
-class LoginActivity : BaseActivity() {
-
+class GuarantorActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        val loanId = intent.getLongExtra(LOAN_ID, -1)
         enableEdgeToEdge()
         setContent {
             MifosMobileTheme {
                 val navController = rememberNavController()
-                RootNavGraph(
-                    startDestination = AuthenticationRoute.AUTH_NAVIGATION_ROUTE,
+                SetUpGuarantorNavGraph(
+                    startDestination = GuarantorScreen.GuarantorList.route,
                     navController = navController,
+                    loanId = loanId,
+                    navigateBack = { finish() }
                 )
             }
         }
     }
 }
+
