+++ app/src/main/java/org/mifos/mobile/ui/activities/PassCodeActivity.kt
@@ -1,6 +1,7 @@
 package org.mifos.mobile.ui.activities
 
 import android.Manifest
+import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import android.provider.Settings
