+++ core/common/src/main/java/org/mifos/mobile/core/common/Constants.kt
@@ -69,11 +69,4 @@ object Constants {
     const val LICENSE_LINK = "https://github.com/openMF/mifos-mobile/blob/development/LICENSE.md"
     const val SOURCE_CODE_LINK = "https://github.com/openMF/mifos-mobile"
     const val WEBSITE_LINK = "https://openmf.github.io/mobileapps.github.io/"
-
-    const val GUARANTOR_LIST_SCREEN_ROUTE_BASE = "guarantor_list_screen"
-    const val GUARANTOR_DETAIL_SCREEN_ROUTE_BASE = "guarantor_detail_screen"
-    const val GUARANTOR_ADD_SCREEN_ROUTE_BASE = "guarantor_add_screen"
-
-    const val REGISTRATION_SCREEN_ROUTE_BASE = "registration_screen"
-    const val REGISTRATION_VERIFICATION_SCREEN_ROUTE_BASE = "registration_verification_screen"
 }
