+++ feature/auth/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    alias(libs.plugins.mifos.android.feature)
+    alias(libs.plugins.mifos.android.library.compose)
+}
+
+android {
+    namespace = "org.mifos.mobile.feature.auth"
+}
+
+dependencies {
+    implementation(projects.ui)
+    implementation(projects.core.common)
+    implementation(projects.core.model)
+    implementation(projects.core.data)
+
+
+    implementation("com.github.ParveshSandila:CountryCodeChooser:1.0")
+
+    testImplementation(libs.junit)
+    androidTestImplementation(libs.androidx.test.ext.junit)
+    androidTestImplementation(libs.espresso.core)
+}
