+++ feature/auth/src/main/java/org/mifos/mobile/feature/auth/navigation/AuthenticationRoute.kt
@@ -1,0 +1,8 @@
+package org.mifos.mobile.feature.auth.navigation
+
+object AuthenticationRoute {
+    const val AUTH_NAVIGATION_ROUTE = "auth_route"
+    const val LOGIN_SCREEN_ROUTE = "login_screen"
+    const val REGISTRATION_SCREEN_ROUTE = "registration_screen"
+    const val REGISTRATION_VERIFICATION_SCREEN_ROUTE = "registration_verification_screen"
+}
