+++ feature/auth/src/main/res/drawable/ic_lock_black_24dp.xml
