+++ feature/auth/src/main/res/drawable/ic_person_black_24dp.xml
