+++ feature/auth/src/main/res/drawable/mifos_logo.png
