--- feature/login/src/main/java/org/mifos/mobile/feature/login/utils/LoginState.kt
@@ -1,9 +1,0 @@
-package org.mifos.mobile.feature.login.utils
-
-sealed class LoginState {
-    object Initial : LoginState()
-    object LoginSuccess : LoginState()
-    object Loading : LoginState()
-    object Error : LoginState()
-    data class LoadClientSuccess(val clientName: String?) : LoginState()
-}
