--- feature/registration/src/main/java/org/mifos/mobile/feature/registration/utils/RegistrationState.kt
@@ -1,8 +1,0 @@
-package org.mifos.mobile.feature.registration.utils
-
-sealed class RegistrationState {
-    data class Error(val exception: Int) : RegistrationState()
-    object Success : RegistrationState()
-    object Loading : RegistrationState()
-    object Initial: RegistrationState()
-}
