+++ settings.gradle.kts
@@ -30,9 +30,7 @@ include(":core:data")
 include(":core:network")
 include(":core:datastore")
 include(":feature:loan")
-include(":feature:login")
 include(":feature:beneficiary")
-include(":feature:registration")
 include(":feature:savings")
 include(":feature:qr")
 include(":feature:transfer-process")
@@ -45,4 +43,5 @@ include(":feature:notification")
 include(":feature:location")
 include(":feature:about")
 include(":feature:settings")
+include(":feature:auth")
 include(":feature:update-password")
